Code/requirements.txt
logging==0.5.1.2
spidev==3.5
Pillow==8.4.0
requests==2.26.0
paho-mqtt==1.6.1
adafruit-circuitpython-ads1x15==2.3.3
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-board==1.0.0
flask==2.0.2
waitress==2.1.0